Add virtual intelligence and user interface for selecting player type
To begin with just adding the dumbest of playstyles for the virtual
intelligence - playing completely randomly.

I'm not a huge fan of my code in the configureGame, setNumberOfPlayers and
configurePlayer functions here, especially the way that they try to keep the
UI and javascript data in sync with each other, but it works for now.