Add the original atoms music (in mp3 format)
The looping is not perfectly gapless, but this seems to be somewhat of a
limitation of current HTML5 audio element support. Similarly there is also
a noticable gap between the intro ending and the main music loop starting.

It may be that this can be improved upon, but for now it is at least good
enough.

Resolves: #1